Remove scavnav and put functions in scavbot
