+++ scav-hunt/config.py
@@ -1,0 +1,6 @@
+params = {
+    'radius': 200,
+    'h_spd': 400,
+    'm_spd': 200,
+    'l_spd': 30
+}
